organize components and update imports
