+++ src/components/home/index.ts
@@ -2,3 +2,4 @@ export * from "./about";
 export * from "./experience";
 export * from "./hero";
 export * from "./projects";
+export * from "./scroll-arrow";
