+++ src/components/home/projects.tsx
@@ -4,7 +4,7 @@ import * as React from "react";
 import projects from "~/data/projects.json";
 
 import Link from "next/link";
-import { Icons } from "~/components";
+import { Icons } from "~/components/icons";
 import {
   Card,
   CardContent,
