+++ src/components/paste/index.ts
@@ -1,0 +1,2 @@
+export * from "./spinner";
+export * from "./text-editor";
