+++ src/components/paste/spinner.tsx
