+++ src/components/paste/text-editor.tsx
