+++ src/components/site/index.ts
@@ -1,9 +1,6 @@
 export * from "./animation-wrapper";
 export * from "./footer";
-export * from "./icons";
 export * from "./menu";
 export * from "./social-icons";
-export * from "./spinner";
-export * from "./text-editor";
 export * from "./theme-provider";
 export * from "./theme-toggle";
